Dsweet/lfg (#32)
* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* o

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

* ok

---------

+++ sampling/sobol_indices.py
@@ -12,11 +12,11 @@ def calculate_sobol_indices_np(x: np.ndarray, y: np.ndarray) -> np.ndarray:
     assert y.ndim == 1
     assert y.shape[0] == n
     if n < 9:
-        return np.ones(d, dtype=float)
+        return np.ones(d, dtype=x.dtype)
     mu = np.mean(y)
     vy = np.var(y, ddof=0)
     if not np.isfinite(vy) or vy <= 0.0:
-        return np.ones(d, dtype=float)
+        return np.ones(d, dtype=x.dtype)
     B = 10 if n >= 30 else 3
     order = np.argsort(x, axis=0, kind="mergesort")
     ranks = np.empty_like(order)
@@ -24,7 +24,7 @@ def calculate_sobol_indices_np(x: np.ndarray, y: np.ndarray) -> np.ndarray:
     col_idx = np.arange(d)[None, :]
     ranks[order, col_idx] = row_idx
     idx = (ranks * B) // n
-    oh = np.eye(B, dtype=float)[idx]
+    oh = np.eye(B, dtype=x.dtype)[idx]
     counts = oh.sum(axis=0)
     sums = (oh * y[:, None, None]).sum(axis=0)
     mu_b = np.divide(sums, counts, out=np.zeros_like(sums), where=counts > 0)
@@ -32,4 +32,4 @@ def calculate_sobol_indices_np(x: np.ndarray, y: np.ndarray) -> np.ndarray:
     diff = mu_b - mu
     S = (p_b * (diff * diff)).sum(axis=1) / vy
     # S = np.clip(S, 0.0, 1.0)
-    return S
+    return S.astype(x.dtype, copy=False)
